requirements.txt
absl-py==0.9.0
astor==0.8.1
atari-py==0.2.6
backcall==0.2.0
cloudpickle==1.2.2
cycler==0.10.0
decorator==4.4.2
Deprecated==1.2.7
future==0.18.2
gast==0.2.2
google-pasta==0.1.8
grpcio==1.27.2
gym==0.16.0
h5py==2.10.0
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.2.1
matplotlib==3.4.1
numpy==1.18.1
opencv-python==4.2.0.32
opt-einsum==3.1.0
pandas==1.0.1
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.18
protobuf==3.11.3
ptyprocess==0.7.0
pyglet==1.5.0
Pygments==2.8.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
pyzmq==22.0.3
scikit-learn==0.24.1
scipy==1.4.1
six==1.14.0
stable-baselines==2.9.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
threadpoolctl==2.1.0
torch==1.4.0
tornado==6.1
traitlets==5.0.5
wcwidth==0.2.5
Werkzeug==1.0.0
wrapt==1.12.0
-e ./Learning/gym_simalphagarden/agenv/
-e ./Learning/gym_simalphagarden/wrapper/
-e ./Simulator